requirements.txt
bitcoin==1.1.42
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
marshmallow-sqlalchemy==0.23.1
Flask-Migrate==2.5.3
Flask-HTTPAuth==4.1.0
flask-marshmallow==0.13.0
psycopg2-binary==2.8.5
pytest==5.4.3
